fix(build): update dependencies for golang 1.23
Running the tests using the old configuration would raise the following
error:

  link: golang.org/x/net/internal/socket: invalid reference to syscall.recvmsg

This is because of a change to golang.org/x/net/internal/socket that
added a forbidden //go:linkname to syscall.recvmesg and first become an
issue with golang 1.23.

Luckily, updating the modules, which bumps the project to use
golang.org/x/net v0.30.0 seems to resolve this issue.

DCO-1.1 Signed-off-by: Patrick Wagstrom <160672+pridkett@users.noreply.github.com>