feat(types): add CNAME, MX, SRV records
This adds in code and tests to support CNAME, MX, and SRV records. In
the case of MX and SRV records, the input should just be space separated
fields as they'd normally appear in a DNS query - this is because the
pdnsql schema doesn't support things like weights and ports.

DCO-1.1 Signed-off-by: Patrick Wagstrom <160672+pridkett@users.noreply.github.com>